Remove "Canvas.prefab" from Rimaethon/Music
The "Canvas.prefab" file in the Rimaethon/Music directory has been removed. It seemed to be a large prefab file that was not being used within the current work scope, resulting in unnecessary bloat within the repository. After checking for references and usages within the project and finding none, it was deemed safe for removal. Cleaning up such unused files can speed up builds and reduce the overall project size.