"Add AudioSyncColor functionality in MusicSync module"
This commit introduces a new file 'AudioSyncColor.cs' into the Simple Audio Visualizer project, specifically within the MusicSync module. This new class, 'AudioSyncColor', inherits from 'AudioPeakListener' and is responsible for controlling the color of an image component based on the amplitude of the audio. As the amplitude peaks, the color of the image transitions from yellow to red. This feature was added to enhance the visual experience of the audio visualization.