Update component and scale for 5-Side Diamond prefab
The component associated with the 5-Side Diamond prefab was updated to rectify an issue with its proportions. The previous component made the item disproportionately small. By updating the component and adjusting the m_LocalScale from 0.21 to 1, the object now appears at its correct size within the game. Additionally, the script associated with the component has been updated to fix additional related bugs.