Enable GameScene and slim down font data
Enabled the GameScene in the project build settings to include it in the final build. The excessive and redundant information present in the Electronic Highway Sign SDF.asset file was cleaned up to reduce memory usage and disk space.

This was achieved by emptying the Glyph table, Character table, Used Glyph Rects, and setting the width and height to zero, effectively making the asset lightweight. Dashboard performance should now improve due to reduced game asset size and possibly improved game scene load times with GameScene enabled in the build settings.