Remove Color Gradient Presets and UI button prefab
Deleted unnecessary 'Color Gradient Presets' directory which was not being used hence cleaning up the project directory. Also removed 'Content 1' prefab from 'Dark UI' as it was result of a previous experiment and is no longer needed. Changes are part of the effort to clean up and better organise the project.