Remove plus icon and Poppins-Bold font from FreeButtonSet
This commit removes the plus icons of various resolutions (32x32, 64x64, 128x128, and 256x256) from the FreeButtonSet since they are no longer used in the project. Additionally, it removes the Poppins-Bold font from the same set, as it had been replaced by another font in a previous update. Removing these resources results in a more efficient use of storage and loading times.